Fiz putnbr functions and rm ft_numlen

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: aarias-d <aarias-d@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/31 21:19:43 by aarias-d          #+#    #+#             */
-/*   Updated: 2025/06/08 19:36:36 by aarias-d         ###   ########.fr       */
+/*   Updated: 2025/06/08 20:04:02 by aarias-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int	ft_validif(char c, va_list ap)
 	else if (c == 'x' || c == 'X')
 		len += ft_printhexadecimal(va_arg(ap, unsigned int ), c);
 	else if (c == 'd' || c == 'i')
-		len += ft_putnbr_fd(va_arg(ap, long), 1);
+		len += ft_putnbr_fd(va_arg(ap, int), 1);
 	else if (c == '%')
 		len += ft_putchar_fd('%', 1);
 	else if (c == 'u')
